Fixed build and added missing vendor entries.
I had to add a different origin for the three roughtime.googlesource.com libraries, as
I could not make it work with govendor fetch or go get for the life of me.
Added roughtime to the makefile and fixed a missing call in the time server to use our scion util library.