Add go.mod, update quic-go, replace docker/term
Create a go.mod file, now that one is available in
scionproto/scion (and netsec-ethz/netsec-scion).

This replaces govendor.

Update quic-go to v0.11.0 -- this is the next, proper release after the
version currently used in scionproto (fd7246d7ed6e).
NOTE: 'git describe' (which is used for the version string that go
modules uses) describes fd7246d7ed6e as v0.7.0-..., which is quite
wrong because the releases are tagged in a branch! It is actually
halfway between v0.10 and v0.11.
NOTE: previously, we vendored v0.8.0 which is just even older than what
scionproto uses.

Replace docker/term with x/crypto/ssh/terminal; this provides identical
functionality & the docker library causes issues with go modules.