Update to scion master (WIP) and wrap snet usage
Add wrapper for snet initialisation and connection establishment.
This wrapper takes care of determining the local address, so we don't
have to specify this when invoking the applications.
Also, it hides away the choice of the dispatcher/sciond socket from the
applications by moving this choice to an environment variable.
This will make usage simpler and more uniform accross applications.