Always set a path in Dial, required after scionproto/scion#3571
Previously, snet would resolve a path every in every write to an
snet.Conn if no path was specified as part of the destination address.
This has been changed in scionproto/scion#3571, now a path is required
if the destination is not in the same AS.
This has some implications on long lived connections and will require
a path management layer on top to make this work again.