Update to scion master and wrap snet usage  (#119)
Add wrapper for snet initialisation and connection establishment, called `appnet`.
This wrapper takes care of determining the local address, so we don't have to specify this when invoking the applications. It does this, by querying sciond for the local IA and by querying the hosts routing table for the appropriate local address. This works without the previously employed workarounds of having a `localhost` entry in `/etc/hosts` or the `gen/ia` file. 
The wrapper hides away the choice of the dispatcher/sciond socket from the applications by moving this choice to an environment variable -- this will only be necessary in runs on the local-topology 
This will make usage simpler and more uniform across applications.

Also, `appnet.Dial`/`DialAddr` take care of always setting a path to the remote address, using the (new) helper function `appnet.SetDefaultPath(addr)`. This is required after scionproto/scion#3571. Previously, snet would resolve a path every in every write to an snet.Conn if no path was specified as part of the destination address. Now a path needs to be set on the address if the destination is not in the same AS. Note: this has some implications on long lived connections and will later require a path management layer on top to make this work again.

Additionally, using the `appnet.Dial` and `appquic.Dial` will resolve hostnames, so that now all apps uniformly support using hostnames as arguments.

Note: a _copy_ of this `appnet` wrapper code will also be used in RAINS (https://github.com/netsec-ethz/rains/pull/231), for the SCION transport in rainsd, rdig and the library. The duplication is obviously undesirable, but the library in scion-apps wants to make use of RAINS to resolve names, so there does not seem to be a way around this without having a third, shared library -- having a shared library just for this would seem to be overly complicated for this.

Also in this PR:

* collate examples in `_examples/` subdirectory, following "standard" golang directory layout. Only keep one client/server pair for shttp examples
* rename lib/ to pkg/ as first step towards better project layout
* helloworld sample application simplified, making use of the `appnet` wrapper. Add a server mode  that prints the message sent to it
* improve Makefile (all targets PHONY, build from root directory to get proper paths in error msg)
* add `make test` target. Disable failing internal tests copied from `bat` clone.
* add `make lint` target using `golangci-lint`. Configured to ignore most of `bat` internals. Also `webapp` and `bwtester` have too many linting to be fixed right now. 
* fix a ton of small lint issues.
* set up CircleCI to run tests and lint

* Fix & combine shttp examples, improve/fix shttp URL mangling

Now shttp offers a helper function MangleSCIONAddrURL that can be used
as a filter to mangle SCION addresses in URLs, before passing them to
URL.Parse etc.
Use this in bat and the example client code to allow using SCION
addresses directly.

* Reorganize reading /etc/hosts to fix tests

Reorganize reading /etc/hosts as this was broken in test set up (would
read both the actual /etc/hosts and the test file).
In the process, change GetNamesByAddress signature take address type not
string (fixes canonicalization problem with string only comparison).