Add path filtering back, fix initialization, update tests
Re-implement path filtering by using pathpol.Filter.

Removed defaultPathSelector as the behaviour is virtually identical with
staticPathSelector now.

Separate path quering from path-selector initialization. This allows
for simpler testing of the path selection logic.
Initialize one path-selector instance per destination ISD-AS. This is
overkill for the current usage; it's clear however that a proper
PacketConn would need to do this, and implementing it is not much more
work than documenting a weird restriction. This would also allow to
eventually use this on the quic _server_.

Update / redo tests for static and round-robin path selectors.