Update to quic-go v0.15.5, adapt usage
Adapt usage of various APIs:
- pass context (`context.Background()`, for now) in various Accept, Listen etc.
- replace `session.Close()` with `session.CloseWithError`
- module `h2quic` is now `http3`
- `http3` now uses `EarlySession` instead of Session; add corresponding `DialEarly` API to `appquic`
- set now required `NextProtos` where needed

At the same time, slightly reorganize the usage of the `tls.Config`
parameter for the `appquic.Dial` functions. As `NextProtos` is now required,
pass this explicitly for every usage.
As a related cleanup, don't automatically set `InsecureSkipVerify` and
self-signed server certificates in the `appquic` library code anymore;
moved to the applications, to make this hack a bit more apparent.