Replace SCMPHandler with higher level RevocationHandler
Using the higher level handler simplifies the code a bit.
This is now mainly for the data connection, but the same revocation
handler is also used by the monitoring connection for receiving echo
replies.

Also, remove the `keyedRevocation`, i.e. revocation with a key
identifying the path over which the revocation arrived. This concept
seemed wrong, as a revocation can affect multiple paths.
As a result, the custom path fingerprinting is no longer needed.