skip: bind address flag, serve skip.pac via http
Add --bind address for local address. Uses kingpin to parse flags simply
because this library is already in use in this repo.

The skip.pac file is now interpreted as a template so that the local
address can be filled in and served on http (instead of previously just
installing it with a file://-URL).
For this, we use the new "go:embed" feature to embed the template file
into the binary (hence the update to go-1.16).

Configuring the skip.pac file with an http://-URL appears to make this
work in chrome :tada:.

Also:
- update to go-1.16 (for go:embed, see above)
- update golangci-lint version in Makefile
- go mod tidy