shttp, skip: use net/http instead of quic-go/http3, support tunneling
The "default" HTTP package for HTTP over SCION, shttp, now uses HTTP/1
or HTTP/2 based on the net/http standard library.
This allows our applications to talk to "normal" (i.e. non-HTTP/3)
web servers / clients when suitably proxied/tunneled.
Also, this now supports http://, which is useful for quickly running the
demo applications without nasty tricks or having to hassle with TLS
setup.

The existing library for HTTP/3 over SCION is moved to a separate
package shttp3, overhauled and reduced to the bare necessities.
This will need some additional love to make it useful (and used).
Normally HTTP/3 is using the same port as HTTPS but on UDP instead of
TCP -- we run both over UDP so we'll have to come up with something to
disambiguate.

Implement CONNECT in skip to support HTTPS. Change the way that SCION
hosts are identified, as the `.scion`-TLD approach is incompatible with
HTTPS (wrong name for certificates). Instead, it just looks at the list
of hosts for which a SCION address is configured in `/etc/hosts` or
`/etc/scion/hosts`.

Add a web gateway that acts as a proxy/gateway to allow mirroring some
content on the TCP/IP web into the SCION based web.