appquic: silence quic-go warnings
quic-go wants to set the read buffer size for the Conn, but snet and the
dispatcher setup does not support this.
Faking the operations to set the read buffer size does not work, as
quic-go also checks directly with a system call whether the operation
was successful.
As a hacky workaround, we simply silence the logger during the quic.Dial
and quic.Listen calls.