pan: add single stream quic utilities
Using a single quic stream is common in our demo applications.
Add a utility function for this, extracted from previous internal
implementation in shttp.
Compared to previous version in shttp, the listener is now more robust.
Previously, if a remote session was opened without ever opening a
stream, this would have blocked the listener Accept indefinitely.