pan: policy-based path-aware SCION application library
This is a first version of the pan "Path-Aware Networking" library.

The pan library represents the path selection process for SCION
connections by Policies, which filter and order the paths based on the
information included in the path metadata, and Selectors, which
dynamically select the final path for each packet sent.
This gives a relatively simple interface for applications that directly
translates to an application user interface (e.g. command line
interface), and connections can automatically failover in case of SCMP
errors messages. The pan library can do the work for querying and
updating path information in the background.
A small number of default Policies and Selectors are built-in. In the
future, we might add more of these out of the box, but the idea is that
applications can also define custom Policies and Selectors.
See the package documentation for more information.

Currently uses inet.af/netaddr to represent IP addresses. This will be
replaced by the very similar net/netip once this becomes available,
expectedly in go-1.18.
Other than this, pan attempts to not expose external libraries in its
API and provide a self-contained interface.

All applications in scion-apps have been converted to pan, but these
changes will be committed separately.

Known issues in this first version:
- the DefaultReplySelector, i.e. the path selector on the listening
  side, is susceptible to session hijacking by spoofing source
  address.
- the stats DB is intentionally kept private as this requires more
  thought. For now, applications need to store their own stats separate
  from pans internal store.
- internal state of DefaultReplySelector and stats DB are never cleaned
  up and can grow without bounds.
- far too many globals

Add package quicutil:
- single stream quic utilities
  Using a single quic stream is common in our demo applications.
  Add a utility function for this, extracted from previous internal
  implementation in shttp.
  Incompatible to the previous version in shttp. Now either side can
  start sending (by using two unidirectional streams), plus it can now
  reliably close. The listener is also more robust; previously, if a
  remote session was opened without ever opening a stream, this would
  have blocked the listener Accept indefinitely.
- generate self-signed certificate (previously in appquic)

Updated example code:
- extend helloworld example and port to pan.
- add helloquic example, using pan

Misc:
- integration: use addr= instead of only ia= as "ready signal"
- remove appnet/appquic from main README as these packages are slated for removal.