Remove appquic, add quicutil for self-signed certs and single stream
Replaced remaining usages of appquic with pan.
Put the useful "generate self-signed certificate" thing into a new
package, quicutil.

Single stream quic "protocol" was first located in pan. Moved to
quicutil now as it's independent and more of a utility for the "TCP-y"
demo applications.
Also, important (and incompatible) fixes for this single stream
protocol. Now either side can start sending, plus it can now reliably
close. This stuff is delicate, it should have lots of tests -- but i
don't quite know how to do this best.

Updated various usages of the "single stream" protocol.
Removed the long list of supported protocols in netcat again -- that was
a bit of a brain fart. The delicate handling of the quic streams is
enough of a protocol to warrant having a dedicated identifier.

Updated netcat QUIC integration tests (dont need the -b now!)

Fixed netcat UDP closing (panic, write on closed channel).