Adapted after adding drkey support to colibri.
DRKey is no longer optional, the local topologies include
configuration for DRKey.
The local topologies also include the delegation entries
for the scion daemon for the two protocols
"colibri" and "piskes".
Updated the local address of the scion daemons for 111 and 112.
Added a timeout to connect to the scion daemon.