Remove camerapp (#208)
Remove camerapp. We've not had a running instance for some time. The
image-retrieval was built as a custom UDP-based ~reliable protocol. With
the libraries we have now, it would definitely make much more sense to
build this as a rather simple webserver instead (or at least, use quic).

Remove camerapp functionality from webapp.