bwtester: convert to pan
Add default path policy options to command line for bwtestclient.

For the data channel, we want to ensure to use the path used for the
control channel and stick to this, i.e. no more fail-over once
connection is established. On the client side, we use the Pin policy to,
well, pin the path for the data channel. On the server side, we
initialize a default reply path selector with the path known from the
control channel. With this approach, we stick to the original forwarding
path, but the path can still be refreshed if it expires (however
unlikely that is, given the short max duration of tests). We could
perhaps go a bit further and use dumb selectors that always return the
fixed, initial path.