skip: add API to set path policy (#223)
Add API `/setPolicy` endpoint to set a path policy in the form of a json-encoded "access control list" (list of hop-predicates).
This will be used by the browser extension "front end" to configure the policy (only filter ISDs in a first version).

Known issue: currently, when the ACL is unmarshalled from JSON it does not check for a defined default. Policies without a default can panic during evaluation. This should be fixed upstream (e.g. add validation during unmarshal analogous to the processing in NewACL function). It's not a big problem for skip though (assume that input is benign and crashing is no big deal).

Also, add an API to retrieve the host names, also for use in the "front end" to configure the proxy usage (analogous to the use in the skip.pac configuration file).