Update to latest netsec-ethz/scion:scionlab, hellodrkey to DRKey v2 (#226)
Update to latest netsec-ethz/scion:scionlab, rewrite hellodrkey for DRKey v2

The hellodrkey example now uses grpc directly, which is not used in the rest of
the libraries. Thus it makes sense to have a separate go module for the
examples. This also allows us to remove the replace directive for
scionproto/scion in the top-level module, as netsec-ethz/scion specific code is
only used in the examples (drkey).
The go build commands for the examples sub-module need to be invoked from
within the corresponding directory. Update the Makefile accordingly.